requirements.txt
Django==1.9
django-filebrowser==3.7.2
gunicorn==19.6.0
Pillow==3.4.2
